Pull "stop" to Ruby and set markers in Ruby
I want to hide the profiler marker implementation from the result object
and the Gecko output object.  It should be the profiler's responsibility
to reconcile internal events.  For example, the profiler should convert
GC start / stop events in to "intervals" rather than leaking that
information in to the result / Gecko object