Move marker reconciliation to the profile object
This commit pushes marker reconciliation to the profile object.  The
result object just holds data, and the FF output object just converts
the data to gecko format.  Now only the profiler knows about how to
reconcile GC events.

Note: we may want to push marker phase and gc enter / exit collection in
to the C++ extension.  Since enter / exit should happen on the same
thread, it seems we could just record it on the profile?  Then we don't
need to post-process the events