Started threads should be considered "running"
We used to only sample running threads because started threads didn't
have the full VM set yet.  In PR #18, we changed the event we listen
for, and when that event is fired it should be safe to collect samples.
This patch makes "started" threads considered as "running" so we'll get
those samples.