Collect all thread names and improve formatting
Register a tracepoint in Ruby so that we can see threads which exit and
collect their names at that time. For threads which continue to run we
can gather their names at profile finish.

For explicitly named threads we should not further mangle their names
(though I can see value in their source location I'd rather we put that
somewhere else when we do have an explicit name)

For threads we are giving a name by performing inspect this uses
object_id as an identifier instead of parsing out the address. This also
replaces gem_dir with an empty string instead of ...