Just map Kernel after cfunc
Makes the logic simpler. We initially made this a special check at the
start as it would be confused with cfunc.

But I can see there are spans further down in the stack like "Array#each"
that will be detected as "Kernel" so I think it's ok to just treat this
"Kernel" category as all the other ones.

This matches what speedscope does as well.